Extractor: common => scraper!
